pyproject.toml
[tool.poetry]
name = "businesscontrolsystem-fastapi "
version = "0.1.0"
description = ""
authors = ["PavelShaura <gkdrdfffe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
black = "^24.4.2"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.3"
redis = "^5.0.7"
fastapi-mail = "^1.4.1"
ruff = "^0.5.2"
sqlalchemy-utils = "^0.41.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
